agent-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
openai==0.28.1
